Move server to integration folder
